// control-plane/mesh-ui/src/api/openapi/types.gen.ts
export type V1CreatePreAuthKeyRequest = {
  user?: string
  reusable?: boolean
  ephemeral?: boolean
  expiration?: string
  aclTags?: Array<string>
}

export type V1CreatePreAuthKeyResponse = {
  preAuthKey?: V1PreAuthKey
}
// ...
export type V1ExpirePreAuthKeyRequest = {
  id?: string
}

export type V1ExpirePreAuthKeyResponse = {
  [key: string]: unknown
}

export type V1GetNodeResponse = {
  node?: V1Node
}

export type V1GetPolicyResponse = {
  policy?: string
  updatedAt?: string
}

export type V1HealthResponse = {
  databaseConnectivity?: boolean
}

export type V1ListApiKeysResponse = {
  apiKeys?: Array<V1ApiKey>
}

export type V1ListNodesResponse = {
  nodes?: Array<V1Node>
}

export type V1ListPreAuthKeysResponse = {
  preAuthKeys?: Array<V1PreAuthKey>
}
// ...
export type V1PreAuthKey = {
  user?: V1User
  id?: string
  key?: string
  reusable?: boolean
  ephemeral?: boolean
  used?: boolean
  expiration?: string
  createdAt?: string
  aclTags?: Array<string>
}
// ...
export type V1User = {
  id?: string
  name?: string
  createdAt?: string
  displayName?: string
  email?: string
  providerId?: string
  provider?: string
  profilePicUrl?: string
}